pyproject.toml
[project]
name = "data-visualizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.97.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.1.0",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]
